Fixed unittests and changed System manager coe_client to orchestrator structure following the agent's architecture
